\name{blm.likelihood}
\alias{blm.likelihood}

\title{Specify the Likelihood for a Bayesian Linear Model}

\description{Specify the Likelihood for a Bayesian Linear Model}

\usage{
blm.likelihood(type = c("norm", "t"), df = 3)
}

\arguments{
  \item{type}{type}
  \item{df}{df}
}

\keyword{likelihood}
