\name{blm.prior}
\alias{blm.prior}

\title{Specify the Priors for a Bayesian Linear Model}

\description{Specify the Priors for a Bayesian Linear Model}

\usage{
blm.prior(priorBeta = fbprior("nonInformative"),
          priorSigma = fbprior("nonInformative"))
}

\arguments{
  \item{priorBeta}{priorBeta}
  \item{priorSigma}{priorSigma}
}

\keyword{prior}


