\name{blm}
\alias{blm}

\title{Bayesian Linear Model}
\description{Bayesian Linear Model}

\usage{
blm(formula, data, subset, weights, na.action, prior = blm.prior(),
    likelihood = blm.likelihood(), sampler = blm.sampler(),
    contrasts = NULL)
}

\arguments{
  \item{formula}{formula}
  \item{data}{data}
  \item{subset}{subset}
  \item{weights}{weights}
  \item{na.action}{na.action}
  \item{prior}{prior}
  \item{likelihood}{likelihood}
  \item{sampler}{sampler}
  \item{contrasts}{contrasts}
}

\keyword{bayes}


